PES-3231: coding standard - migrate Install/UpgradeData to patches
