--- Packetery/Checkout/Setup/InstallData.php
@@ -1,22 +1,0 @@
-<?php
-
-
-namespace Packetery\Checkout\Setup;
-
-use Magento\Framework\Setup\InstallDataInterface;
-use Magento\Framework\Setup\ModuleContextInterface;
-use Magento\Framework\Setup\ModuleDataSetupInterface;
-
-class InstallData implements InstallDataInterface
-{
-
-    /**
-     * {@inheritdoc}
-     */
-    public function install(
-        ModuleDataSetupInterface $setup,
-        ModuleContextInterface $context
-    ) {
-        //Your install script
-    }
-}
