--- Packetery/Checkout/Setup/UpgradeData.php
@@ -1,166 +1,0 @@
-<?php
-
-namespace Packetery\Checkout\Setup;
-
-use Magento\Framework\Setup\ModuleContextInterface;
-use Magento\Framework\Setup\ModuleDataSetupInterface;
-use Magento\Framework\Setup\UpgradeDataInterface;
-
-class UpgradeData implements UpgradeDataInterface
-{
-    /** @var \Magento\Config\Model\Config\Factory */
-    private $configFactory;
-
-    /** @var \Packetery\Checkout\Model\ResourceModel\Pricingrule\CollectionFactory */
-    private $pricingRuleCollectionFactory;
-
-    /** @var \Magento\Framework\App\Config\Storage\WriterInterface */
-    private $configWriter;
-
-    /** @var \Magento\Store\Model\StoreManagerInterface */
-    private $storeManager;
-
-    /**
-     * @param \Magento\Config\Model\Config\Factory $configFactory
-     * @param \Packetery\Checkout\Model\ResourceModel\Pricingrule\CollectionFactory $pricingRuleCollectionFactory
-     * @param \Magento\Framework\App\Config\Storage\WriterInterface $configWriter
-     * @param \Magento\Store\Model\StoreManagerInterface $storeManager
-     */
-    public function __construct(
-        \Magento\Config\Model\Config\Factory $configFactory,
-        \Packetery\Checkout\Model\ResourceModel\Pricingrule\CollectionFactory $pricingRuleCollectionFactory,
-        \Magento\Framework\App\Config\Storage\WriterInterface $configWriter,
-        \Magento\Store\Model\StoreManagerInterface $storeManager
-    ) {
-        $this->configFactory = $configFactory;
-        $this->pricingRuleCollectionFactory = $pricingRuleCollectionFactory;
-        $this->configWriter = $configWriter;
-        $this->storeManager = $storeManager;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function upgrade(
-        ModuleDataSetupInterface $setup,
-        ModuleContextInterface $context
-    ): void {
-        if (version_compare($context->getVersion(), "2.1.0", "<")) {
-            $configModel = $this->configFactory->create();
-            $configModel->setDataByPath('carriers/packetery/sallowspecific', 0); // config option UI was removed
-            $configModel->save();
-        }
-
-        if (version_compare($context->getVersion(), '2.4.0', '<')) {
-            $packeteryOrderTable = $setup->getTable('packetery_order');
-            $salesOrderTable = $setup->getTable('sales_order');
-            $salesOrderAddressTable = $setup->getTable('sales_order_address');
-            $setup->getConnection()->query("
-                UPDATE `$packeteryOrderTable`
-                JOIN `$salesOrderTable` ON `$salesOrderTable`.`increment_id` = $packeteryOrderTable.`order_number`
-                JOIN `$salesOrderAddressTable` ON `$salesOrderTable`.`shipping_address_id` IS NOT NULL AND `$salesOrderAddressTable`.`entity_id` = `$salesOrderTable`.`shipping_address_id`
-                SET `$packeteryOrderTable`.`recipient_country_id` = `$salesOrderAddressTable`.`country_id`
-                WHERE `$packeteryOrderTable`.`recipient_country_id` IS NULL
-            ");
-
-            // Static Vendor groups snapshot for given module version 2.3.0
-            $vendorGroupsMapping = [
-                'CZ' => '["zpoint","zbox"]',
-                'SK' => '["zpoint","zbox"]',
-                'HU' => '["zpoint","zbox"]',
-                'RO' => '["zpoint","zbox"]',
-            ];
-
-            $countries = $this->pricingRuleCollectionFactory->create();
-            $countries
-                ->getSelect()
-                ->reset(\Magento\Framework\DB\Select::COLUMNS)
-                ->columns('country_id')
-                ->group('country_id');
-            $pricingCountries = $countries->getColumnValues('country_id');
-
-            foreach ($pricingCountries as $countryId) {
-                $vendorGroupsValue = $vendorGroupsMapping[$countryId] ?? null;
-                if ($vendorGroupsValue === null) {
-                    continue;
-                }
-
-                $setup->getConnection()->update(
-                    $setup->getTable('packetery_pricing_rule'),
-                    [
-                        'vendor_groups' => $vendorGroupsValue,
-                    ],
-                    [
-                        '`method` = ?' => 'pickupPointDelivery',
-                        '`carrier_code` = ?' => 'packetery',
-                        '`country_id` = ?' => $countryId,
-                        new \Zend_Db_Expr('`carrier_id` IS NULL'),
-                        new \Zend_Db_Expr('`vendor_groups` IS NULL'),
-                    ],
-                );
-            }
-
-            $pricingRuleTable = $setup->getTable('packetery_pricing_rule');
-            $carrierTable = $setup->getTable('packetery_carrier');
-            $connection = $setup->getConnection();
-            if ($connection->isTableExists($carrierTable) && $connection->tableColumnExists($carrierTable, 'carrier_name')) {
-                $connection->query("
-                    UPDATE `$pricingRuleTable` pr
-                    INNER JOIN `$carrierTable` c ON c.carrier_id = pr.carrier_id AND pr.carrier_id IS NOT NULL
-                    SET pr.carrier_name = c.carrier_name
-                    WHERE c.carrier_name IS NOT NULL AND c.carrier_name != ''
-                ");
-                $connection->dropColumn($carrierTable, 'carrier_name');
-            }
-        }
-
-        if (version_compare($context->getVersion(), '2.5.0', '<')) {
-            $this->migrateSenderConfig();
-        }
-    }
-
-    private function migrateSenderConfig(): void
-    {
-        $defaultSender = null;
-        foreach ($this->storeManager->getStores() as $store) {
-            $group = $store->getGroup();
-            if ($group !== null) {
-                $groupCode = $group->getCode();
-                if ($groupCode !== null && $groupCode !== '') {
-                    if ($defaultSender === null) {
-                        $defaultSender = $groupCode;
-                    }
-                    $this->configWriter->save(
-                        \Packetery\Checkout\Model\Carrier\Imp\Packetery\Config::CONFIG_PATH_SENDER,
-                        $groupCode,
-                        \Magento\Store\Model\ScopeInterface::SCOPE_STORE,
-                        (string) $store->getId()
-                    );
-                }
-            }
-        }
-        foreach ($this->storeManager->getWebSites() as $website) {
-            $defaultStore = $website->getDefaultStore();
-            if ($defaultStore !== null) {
-                $group = $defaultStore->getGroup();
-                $groupCode = $group !== null ? $group->getCode() : null;
-                if ($groupCode !== null && $groupCode !== '') {
-                    $this->configWriter->save(
-                        \Packetery\Checkout\Model\Carrier\Imp\Packetery\Config::CONFIG_PATH_SENDER,
-                        $groupCode,
-                        \Magento\Store\Model\ScopeInterface::SCOPE_WEBSITE,
-                        (string) $website->getId()
-                    );
-                }
-            }
-        }
-        if ($defaultSender !== null) {
-            $this->configWriter->save(
-                \Packetery\Checkout\Model\Carrier\Imp\Packetery\Config::CONFIG_PATH_SENDER,
-                $defaultSender,
-                'default',
-                '0'
-            );
-        }
-    }
-}
